requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Data storage
pyarrow>=13.0.0  # for parquet files

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# GUI
gradio>=4.0.0

# Configuration
PyYAML>=6.0.0

# Logging and utilities
tqdm>=4.66.0
coloredlogs>=15.0.0

# Deep learning (Phase 2)
# torch>=2.0.0
# tensorflow>=2.13.0

# Optional: SHAP for feature importance
# shap>=0.42.0
